Combine SetupWorkingDirectory for Linux and Windows
