+++ container/container_windows.go
@@ -3,6 +3,9 @@
 package container
 
 import (
+	"os"
+	"path/filepath"
+
 	"github.com/docker/docker/daemon/execdriver"
 	"github.com/docker/docker/volume"
 	"github.com/docker/engine-api/types/container"
@@ -20,12 +23,6 @@ type Container struct {
 func (container *Container) CreateDaemonEnvironment(linkedEnv []string) []string {
 	// On Windows, nothing to link. Just return the container environment.
 	return container.Config.Env
-}
-
-// SetupWorkingDirectory initializes the container working directory.
-// This is a NOOP In windows.
-func (container *Container) SetupWorkingDirectory() error {
-	return nil
 }
 
 // UnmountIpcMounts unmount Ipc related mounts.
@@ -58,4 +55,16 @@ func (container *Container) UpdateContainer(hostConfig *container.HostConfig) er
 // this is a no-op.
 func appendNetworkMounts(container *Container, volumeMounts []volume.MountPoint) ([]volume.MountPoint, error) {
 	return volumeMounts, nil
+}
+
+// cleanResourcePath cleans a resource path by removing C:\ syntax, and prepares
+// to combine with a volume path
+func cleanResourcePath(path string) string {
+	if len(path) >= 2 {
+		c := path[0]
+		if path[1] == ':' && ('a' <= c && c <= 'z' || 'A' <= c && c <= 'Z') {
+			path = path[2:]
+		}
+	}
+	return filepath.Join(string(os.PathSeparator), path)
 }
